refactor: move the shared diffusion code into diffusion.py
train.py and test.py each carried their own copy of var_func_vp,
var_func_geometric, extract, get_time_schedule, get_sigma_schedule,
Posterior_Coefficients, sample_posterior and sample_from_model -- about 120
duplicated lines defining the sampler that both the training loop and
inference depend on. The two copies are currently byte-identical, so any
change to the diffusion process has to be made twice and silently produces a
train/test mismatch if it is not.

Move them into diffusion.py, together with Diffusion_Coefficients, q_sample
and q_sample_pairs, which only train.py had. The definitions are moved
verbatim; both entry points now import from the single copy.

Behaviour is unchanged: the extracted sources were checked to be identical
between the two files before merging.
