+++ test.py
@@ -7,6 +7,8 @@ import os
 import torch.optim as optim
 import torchvision
 from backbones.ncsnpp_generator_adagn import NCSNpp
+from diffusion import (Posterior_Coefficients, get_time_schedule,
+                       sample_from_model)
 from dataset import CreateDatasetSynthesis
 
 import torch.nn.functional as F
@@ -18,120 +20,6 @@ def psnr(img1, img2):
 
     mse = torch.mean((img1 - img2) ** 2)
     return 20 * torch.log10(img1.max() / torch.sqrt(mse))
-
-        
-#%% Diffusion coefficients 
-def var_func_vp(t, beta_min, beta_max):
-    log_mean_coeff = -0.25 * t ** 2 * (beta_max - beta_min) - 0.5 * t * beta_min
-    var = 1. - torch.exp(2. * log_mean_coeff)
-    return var
-
-def var_func_geometric(t, beta_min, beta_max):
-    return beta_min * ((beta_max / beta_min) ** t)
-
-def extract(input, t, shape):
-    out = torch.gather(input, 0, t)
-    reshape = [shape[0]] + [1] * (len(shape) - 1)
-    out = out.reshape(*reshape)
-
-    return out
-
-def get_time_schedule(args, device):
-    n_timestep = args.num_timesteps
-    eps_small = 1e-3
-    t = np.arange(0, n_timestep + 1, dtype=np.float64)
-    t = t / n_timestep
-    t = torch.from_numpy(t) * (1. - eps_small)  + eps_small
-    return t.to(device)
-
-def get_sigma_schedule(args, device):
-    n_timestep = args.num_timesteps
-    beta_min = args.beta_min
-    beta_max = args.beta_max
-    eps_small = 1e-3
-   
-    t = np.arange(0, n_timestep + 1, dtype=np.float64)
-    t = t / n_timestep
-    t = torch.from_numpy(t) * (1. - eps_small) + eps_small
-    
-    if args.use_geometric:
-        var = var_func_geometric(t, beta_min, beta_max)
-    else:
-        var = var_func_vp(t, beta_min, beta_max)
-    alpha_bars = 1.0 - var
-    betas = 1 - alpha_bars[1:] / alpha_bars[:-1]
-    
-    first = torch.tensor(1e-8)
-    betas = torch.cat((first[None], betas)).to(device)
-    betas = betas.type(torch.float32)
-    sigmas = betas**0.5
-    a_s = torch.sqrt(1-betas)
-    return sigmas, a_s, betas
-
-#%% posterior sampling
-class Posterior_Coefficients():
-    def __init__(self, args, device):
-        
-        _, _, self.betas = get_sigma_schedule(args, device=device)
-        
-        #we don't need the zeros
-        self.betas = self.betas.type(torch.float32)[1:]
-        
-        self.alphas = 1 - self.betas
-        self.alphas_cumprod = torch.cumprod(self.alphas, 0)
-        self.alphas_cumprod_prev = torch.cat(
-                                    (torch.tensor([1.], dtype=torch.float32,device=device), self.alphas_cumprod[:-1]), 0
-                                        )               
-        self.posterior_variance = self.betas * (1 - self.alphas_cumprod_prev) / (1 - self.alphas_cumprod)
-        
-        self.sqrt_alphas_cumprod = torch.sqrt(self.alphas_cumprod)
-        self.sqrt_recip_alphas_cumprod = torch.rsqrt(self.alphas_cumprod)
-        self.sqrt_recipm1_alphas_cumprod = torch.sqrt(1 / self.alphas_cumprod - 1)
-        
-        self.posterior_mean_coef1 = (self.betas * torch.sqrt(self.alphas_cumprod_prev) / (1 - self.alphas_cumprod))
-        self.posterior_mean_coef2 = ((1 - self.alphas_cumprod_prev) * torch.sqrt(self.alphas) / (1 - self.alphas_cumprod))
-        
-        self.posterior_log_variance_clipped = torch.log(self.posterior_variance.clamp(min=1e-20))
-        
-def sample_posterior(coefficients, x_0,x_t, t):
-    
-    def q_posterior(x_0, x_t, t):
-        mean = (
-            extract(coefficients.posterior_mean_coef1, t, x_t.shape) * x_0
-            + extract(coefficients.posterior_mean_coef2, t, x_t.shape) * x_t
-        )
-        var = extract(coefficients.posterior_variance, t, x_t.shape)
-        log_var_clipped = extract(coefficients.posterior_log_variance_clipped, t, x_t.shape)
-        return mean, var, log_var_clipped
-    
-  
-    def p_sample(x_0, x_t, t):
-        mean, _, log_var = q_posterior(x_0, x_t, t)
-        
-        noise = torch.randn_like(x_t)
-        
-        nonzero_mask = (1 - (t == 0).type(torch.float32))
-
-        return mean + nonzero_mask[:,None,None,None] * torch.exp(0.5 * log_var) * noise
-            
-    sample_x_pos = p_sample(x_0, x_t, t)
-    
-    return sample_x_pos
-
-def sample_from_model(coefficients, generator, n_time, x_init, T, opt):
-    x = x_init[:,[0],:]
-    source = x_init[:,[1],:]
-    with torch.no_grad():
-        for i in reversed(range(n_time)):
-            t = torch.full((x.size(0),), i, dtype=torch.int64).to(x.device)
-          
-            t_time = t
-            latent_z = torch.randn(x.size(0), opt.nz, device=x.device)#.to(x.device)
-            x_0 = generator(torch.cat((x,source),axis=1), t_time, latent_z)
-            x_new = sample_posterior(coefficients, x_0[:,[0],:], x, t)
-            x = x_new.detach()
-        
-    return x
 
 
 def load_checkpoint(checkpoint_dir, netG, name_of_network, epoch,device = 'cuda:0'):
